requirements.txt
run streamlit
pandas
numpy
scipy
plotly
openpyxl
streamlit-extras
plotnine